+++ manifest.json
@@ -1,6 +1,6 @@
 {
   "manifest_version": 3,
-  "version": "1.0.2",
+  "version": "1.0.3",
   "name": "Base64 Converter",
   "description": "It helps to encode or decode the dragged text to base64",
